requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.3
Brotli==1.1.0
bs4==0.0.2
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
chardet==5.2.0
charset-normalizer==3.3.2
constantly==23.10.4
contourpy==1.2.1
cryptography==42.0.8
cssselect==1.2.0
cycler==0.12.1
daphne==4.1.2
defusedxml==0.8.0rc2
distro==1.9.0
dj-database-url==0.5.0
Django==5.0.6
django-allauth==0.63.6
django-analytical==3.1.0
django-ckeditor==6.7.1
django-js-asset==2.2.0
django-social-auth==0.7.28
django-storages==1.14.4
docx==0.2.4
Faker==26.0.0
fonttools==4.53.1
frozenlist==1.4.1
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-core==2.4.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.5
lxml==5.2.2
matplotlib==3.9.1
msgpack==1.0.8
multidict==6.0.5
numpy==2.0.0
oauth2==1.9.0.post1
oauthlib==3.2.2
openai==0.28.0
packaging==24.1
pandas==2.2.2
parsel==1.9.1
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.27.2
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.2
python-openid==2.2.5
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2024.1
redis==5.0.7
reportlab==4.2.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.5.1
scipy==1.14.0
service-identity==24.1.0
setuptools==71.1.0
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
soupsieve==2.5
sqlparse==0.5.0
threadpoolctl==3.5.0
tqdm==4.66.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
w3lib==2.2.1
whitenoise==6.7.0
yarl==1.9.4
zope.interface==6.4.post2
